requirements.txt
deepl
pyperclip
python-dotenv
Pillow 
